feature-erms-800
for ERMS-800:
- 500 when no end date was given for pair is fixed
- alternative navigation generation (the method has been moved to a service to avoid static call)
