+++ app/application.properties
@@ -1,7 +1,7 @@
 #Grails Metadata file
-#Mon Jan 21 10:37:09 CET 2019
-app.buildDate=21.01.2019 10\:37
-app.buildNumber=6548
+#Tue Jan 22 14:40:07 CET 2019
+app.buildDate=22.01.2019 14\:40
+app.buildNumber=6550
 app.buildProfile=development
 app.grails.version=2.5.6
 app.name=laser
